Cleaned up LEDStrip class
